- Added `preLoad` function to plugin interface.
`preLoad` is invoked just before a plugin is activated. This gives an oppertunity to the plugin to initialize certain stuff before the activation and the player content(if any) is loaded. `onLoad` function is now invoked after the player content is loaded and ready for DOM access.

Added `onDone` callback parameter to `MC.loadContent` and there-after `MC.activatePlugin` functions.

Refactored `PlayMediaAPI` to invoke `Player.play` on callback of `MC.activatePlugin` for better control flow.